requirements.txt
requests
ffmpeg-python
python-dotenv
groq
pyaudio
keyboard